// src/shared/errors/enum/errorCategory.enum.ts
enum ErrorCategory {
     VALIDATION = "validation",
	AUTHENTICATION = "authentication",
	AUTHORIZATION = "authorization",
	NOT_FOUND = "not_found",
	CONFLICT = "conflict",
	POLICY = "policy",
	WORKFLOW = "workflow",
	INFRASTRUCTURE = "infrastructure",
	SERVER = "server"
}

export { ErrorCategory };